requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
groq==0.12.0
faiss-cpu==1.13.1
sentence-transformers==3.4.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.4
pytest-asyncio==0.24.0
